Checkpoint, simplification branch.
Change from using struct ObjId where we keep
everything in arrays and binary_search for objects
in those arrays, to ObjId being a reference type, and
instead of arrays, objects are kept in circular
doubly linked lists.

Checkpointing to get help with a typing problem with a call
to xnotify_spawn.