Added basic support for drawing motion trails.
Useful for fast things that can move more than
once during a player turn ,so you can see where it
came from.

Disabled because it looks bad.  I can't use gl_line because
I need to taper and alpha-fade the lines to give an idea
of direction.  Since the lines are drawn as the objects move,
can't just create a mesh with it.

Can probably fake it with fl_rect_tc_col and a gradient texture,
but not worth working on right now.