Fix texture bleeding at edges.
Orignal tiles had no border so colors could bleed at the
edges of textures.

Ran into a problem using a tool that extrudes borders out from the
tiles in an image.  TileKit allows you to specify a tile spacing,
but the spacing only happens to the right and bottom of tiles.

I prefer the extruding for sprites, so split out sprites
from maps tiles into a separate png file, and applied
the two different types of borders for each, so TileKit will
be happy with tiles.png, and we can use the extruded border
implementation for sprites.png.

Source images without borders are in src/images, src/images/build
adds borders and copies the results to dist.