Don't store ObjId in linked lists.
That was a dumb idea.  The links made inspecting objects
for debugging a much larger hassle than it needed to be.

Now just keep an array that has nil entries for deleted spots.
Deleted spots are re-used with a naive scan for the first
empty slot.  That can be made smarter with a free list if it
actually ends up being necessary.

Also folded in some enum and gobj changes for equipment.